feature: set up webpack Part-1
Convert stylesheet to scss and import it into the index.js in the client folder. Also import and export the js files into same index.js.
In the server folder, the server.js was also set to serve the dist folder as that would be what is eventually used in production.